// app/requests/getGeoData/getGeoDataSuccess.type.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface GetGeoDataSuccess {
  response: {
    GeoObjectCollection: {
      featureMember: {
        GeoObject: {
          metaDataProperty: {
            GeocoderMetaData: {
              text: string;
              [k: string]: unknown;
            };
            [k: string]: unknown;
          };
          Point: {
            pos: string;
            [k: string]: unknown;
          };
          [k: string]: unknown;
        };
        [k: string]: unknown;
      }[];
      [k: string]: unknown;
    };
    [k: string]: unknown;
  };
  [k: string]: unknown;
}